consensus: Don't add spent coins to the cache
When fetching a coin from a parent cache, don't add it to the child
cache if it's already spent.

This adds an invariable to the coins cache - a coin can't be both FRESH
and spent. This makes sense. If a coin is spent in the child cache and
doesn't exist in the parent cache, then there's no need to keep it in
the child cache.